remove duplicate output; deemphasize timestamps
There are a ton of 'output: xxxx' lines that simply echo
what is in the lines just above. Remove them.

My filter isn't perfect, because (I think) sometimes there's
output on stderr which goes to console but not to the 'output'
variable, but it actually reduces clutter by a great deal.

Signed-off-by: Ed Santiago <ed@edsantiago.com>