try to link to annotated log from github Checks
Github Checks page has a "View more details on Cirrus CI" link
at the bottom. This requires a lot of clickies to get to the
annotated link.

Try to figure it out ourself, by getting the Cirrus task ID
and the test name ("int-this-that").

This fails for APIv2, but most of the time I only care about
int and sys anyway